vm2

+++ src/vm/swap.h
@@ -1,7 +1,9 @@
 #ifndef VM_SWAP_H
 #define VM_SWAP_H
 
-typedef unsigned int slot_t;
+#include <stddef.h>
+
+typedef size_t slot_t;
 
 void swap_init(void);
 slot_t swap_alloc(const void *frame);
